Weather App
PROBLEM STATEMENT:
Create a weather application using Javascript.
DESCRIPTION:
1. Create a User interface with an autocomplete dropdown including a list of
countries/states.
2. An output section showing the result of a search for weather details.
3. Functionality:
a. When a user types or selects a location from the dropdown list and clicks on the
search button, an openWeather API gets called with selected parameters.
b. The output with the current weather details for that location will be displayed.
TECHNOLOGIES USED:
1. HTML
2. CSS
3. JavaScript
4. Openweathermap.org for free API to generate real time weather information.